==> Employee-Attrition-Prediction/Notebooks/01_data-intro.ipynb <==
%pip install -r requirements.txt

#loading the necessary Libraries for the Data(df) Introduction
import pandas as pd 
import numpy as np


#reading Csv File and Data loading using Df 
data_path = (r"C:\Users\ishaan.narayan\Desktop\Ishaan's Workspace\Intern_ML-DL-Gen\Employee-Attrition-Prediction\Data\WA_Fn-UseC_-HR-Employee-Attrition.csv")
df = pd.read_csv(data_path)

print(df.head())


print(df.columns)

print(df.info())

print(df.describe())

print(df.tail())

#checking for the Missing values if any ? 

miss = df.isnull().sum()
print(miss)

#percentage of missing values 
miss_perc = (df.isnull().mean() * 100)
print(miss_perc)

#This shows we have no misisng values yay
