==> Employee-Attrition-Prediction/Notebooks/03_featureEng.ipynb <==
#here we will be considering the oneHotEncoding and Pca and Lda 
# if we find the need to do so as some features might be of no use
# and would increase the Models Complexity.

import pandas as pd
import numpy as np

# importing the Dataset
data_path = (r"C:\Users\ishaan.narayan\Desktop\Ishaan's Workspace\Intern_ML-DL-Gen\Employee-Attrition-Prediction\Data\WA_Fn-UseC_-HR-Employee-Attrition.csv")
df = pd.read_csv(data_path)

df = df.drop(['EmployeeNumber', 'EmployeeCount', 'Over18', 'StandardHours'], axis=1)
print(df.columns)

#using get_dummies for encoding the features of the datset then we will be usin 
# the same features with Xg Boost and Random Forst 

df = pd.get_dummies(df, columns=['BusinessTravel',
                                 'Department',
                                 'EducationField', 
                                 'JobRole',
                                 'MaritalStatus',
                                 'OverTime'], drop_first=True,bool=True)



print(df.columns)

print(df[['BusinessTravel_Travel_Frequently',
          'BusinessTravel_Travel_Rarely',
          'MaritalStatus_Single',
          'MaritalStatus_Married',
          'OverTime_Yes',
          'JobRole_Human Resources',
          'JobRole_Laboratory Technician',
          'JobRole_Manager',
          'JobRole_Manufacturing Director',
          'JobRole_Research Director',
          'JobRole_Research Scientist',
          'JobRole_Sales Executive',
          'JobRole_Sales Representative',    # <-- comma added here
          'Department_Research & Development',
          'Department_Sales',
          'EducationField_Life Sciences',
          'EducationField_Marketing',
          'EducationField_Medical',
          'EducationField_Other',
          'EducationField_Technical Degree'  # <-- comma removed here, 'OverTime' is not needed
         ]].head(10))

#Lets now Export our New Feture Updated Data 
df.to_csv('cleaned_employee_data.csv', index=False)
import os 
print(os.getcwd())

#shifting the data into , DataFolder for better understandin and folder structure maintenance
