==> Employee-Attrition-Prediction/Notebooks/04_Split-GridCV-Rfr.ipynb <==
#instead of loading all at once we will be loading each library when required 
import pandas as pd 
import numpy as np 

df = pd.read_csv(r"C:\Users\ishaan.narayan\Desktop\Ishaan's Workspace\Intern_ML-DL-Gen\Employee-Attrition-Prediction\Data\cleaned_employee_data.csv")
print(df.head())

#loading data 
X = df.drop('Attrition', axis=1)
print(X.shape)

y = df['Attrition']
print(y.shape)

#converting themto TrainTestSplit
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X,y ,random_state=42, test_size=0.3,stratify=y)
print(X_train.shape)
print(X_test.shape)
print(y_test.shape)
print(y_train.shape)



#now i need to find the Best parameters for the RandomForest Classifier 
#using Grid Search For the TASK
#setting up a parameter Grid so that that can be Provided to the GridSearch inb Library 

from sklearn.model_selection import GridSearchCV
my_param_grid = {
    'n_estimators': [100,200,300],
    'max_depth':[None,10,20,3,7]}


#now Importin the Model that we need to use with the GridSeachrdCV 
from sklearn.ensemble import RandomForestClassifier
rf = RandomForestClassifier(random_state=42)
print(rf)
#rf is the object of the randomdomForest 


#now i will be setting up the GridSeachFunction for Finding out the Best set of Prameters 
grid_search = GridSearchCV(
    estimator=rf,
    param_grid=my_param_grid,
    n_jobs=-1,
    scoring='accuracy'
)



#fitting the model with X and y
grid_search.fit(X_train, y_train)
print("Best Parameters:", grid_search.best_params_)
print("Best Cross-Validation Score:", grid_search.best_score_)
